Adiciona a tela de verificação de email e a lógica correspondente, incluindo a nova atividade VerifyEmailActivity. Atualiza o AndroidManifest.xml para registrar a nova atividade e ajusta a navegação na MainActivity. Implementa a lógica de verificação de email no AuthApiService e cria os DTOs necessários para a requisição e resposta. Adiciona novos layouts para a tela de verificação de email e fragmento de exercícios, além de adaptadores para exibir os dados de exercícios.

+++ app/src/main/java/com/ufrj/escalaiv2/dto/PrevisaoAfastamentoResponse.java
@@ -11,7 +11,7 @@ public class PrevisaoAfastamentoResponse {
     private int tempoAfastamentoDias;
 
     @SerializedName("tempo_afastamento_semanas")
-    private int tempoAfastamentoSemanas;
+    private double tempoAfastamentoSemanas;
 
     @SerializedName("confianca")
     private double confianca;
@@ -47,11 +47,11 @@ public class PrevisaoAfastamentoResponse {
         this.tempoAfastamentoDias = tempoAfastamentoDias;
     }
 
-    public int getTempoAfastamentoSemanas() {
+    public double getTempoAfastamentoSemanas() {
         return tempoAfastamentoSemanas;
     }
 
-    public void setTempoAfastamentoSemanas(int tempoAfastamentoSemanas) {
+    public void setTempoAfastamentoSemanas(double tempoAfastamentoSemanas) {
         this.tempoAfastamentoSemanas = tempoAfastamentoSemanas;
     }
